requirements.txt
appdirs==1.4.4
arandr==0.1.11
asgiref==3.6.0
astroid==2.14.2
asttokens==2.4.1
attrs==22.2.0
av==11.0.0
Babel==2.10.3
backcall==0.2.0
beautifulsoup4==4.11.2
beniget==0.4.1
blinker==1.7.0
Brotli==1.0.9
bytecode==0.14.0
certifi==2022.9.24
chardet==5.1.0
charset-normalizer==3.0.1
click==8.1.7
colorama==0.4.6
colorzero==2.0
contourpy==1.0.7
coverage==6.5.0
cryptography==38.0.4
cupshelpers==1.0
cycler==0.11.0
dbus-python==1.3.2
debugpy==1.6.3+git20221103.a2a3328
decorator==5.1.1
dill==0.3.6
distro==1.8.0
docutils==0.19
entrypoints==0.4
executing==2.0.1
Flask==3.0.0
fonttools==4.38.0
fs==2.4.16
gast==0.5.2
gpiozero==2.0
guizero==1.4.0
html5lib==1.1
idna==3.3
importlib-metadata==4.12.0
ipykernel==6.17.0
ipython==8.17.2
ipython_genutils==0.2.0
isort==5.6.4
itsdangerous==2.1.2
jedi==0.19.1
Jinja2==3.1.2
jupyter_client==7.4.9
jupyter_core==4.12.0
kiwisolver==0.0.0
lazy-object-proxy==1.9.0
lgpio==0.2.2.0
libevdev==0.5
logilab-common==1.9.8
lxml==4.9.2
lz4==4.0.2+dfsg
MarkupSafe==2.1.3
matplotlib==3.6.3
matplotlib-inline==0.1.6
mccabe==0.7.0
more-itertools==8.10.0
mpmath==0.0.0
mypy==1.0.1
mypy-extensions==0.4.3
nest-asyncio==1.5.4
nudatus==0.0.5
numpy==1.26.1
oauthlib==3.2.2
olefile==0.46
opencv-python==4.8.1.78
packaging==23.0
paho-mqtt==1.6.1
parso==0.8.3
pexpect==4.8.0
pgzero==1.2
picamera2==0.3.14
pickleshare==0.7.5
pidng==4.0.9
piexif==1.1.3
pigpio==1.78
Pillow==10.1.0
platformdirs==2.6.0
ply==3.11
prompt-toolkit==3.0.39
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycairo==1.20.1
pycodestyle==2.10.0
pycups==2.0.1
pydevd==2.9.5
pyflakes==2.5.0
pygame==2.1.2
Pygments==2.16.1
PyGObject==3.42.2
pyinotify==0.9.6
PyJWT==2.6.0
pylint==2.16.2
PyOpenGL==3.1.6
pyOpenSSL==23.0.0
pyparsing==3.0.9
PyQt5==5.15.9
PyQt5-sip==12.11.1
PyQtWebEngine==5.15.6
pyserial==3.5
pysmbc==1.0.23
python-apt==2.6.0
python-dateutil==2.8.2
python-dotenv==0.21.0
python-prctl==1.8.1
pythran==0.11.0
pytz==2022.7.1
pyudev==0.24.0
pyzmq==24.0.1
QScintilla==2.13.3
qtconsole==5.4.0
QtPy==2.3.0
reportlab==3.6.12
requests==2.28.1
requests-oauthlib==1.3.0
responses==0.18.0
roman==3.3
RPi.GPIO==0.7.1
rpimotorlib==3.2
RTIMULib==7.2.1
scipy==1.10.1
semver==2.10.2
Send2Trash==1.8.1b0
sense-hat==2.4.0
simplejpeg==1.7.2
simplejson==3.18.3
six==1.16.0
smbus2==0.4.2
soupsieve==2.3.2
spidev==3.5
ssh-import-id==5.10
stack-data==0.6.3
sympy==1.11.1
thonny==4.1.3
toml==0.10.2
tomlkit==0.11.7
tornado==6.2
traitlets==5.13.0
twython==3.8.2
typing_extensions==4.4.0
uflash==1.2.4
ufoLib2==0.14.0
urllib3==1.26.12
v4l2-python3==0.3.2
wcwidth==0.2.9
webencodings==0.5.1
Werkzeug==3.0.1
wrapt==1.14.1
wxPython==4.2.0
zipp==1.0.0
